150921.17.3 -- restructuring the folders & action creators
