+++ src/store/actions/burgerBuilder.js
@@ -1,0 +1,15 @@
+import * as actionTypes from './actionTypes';
+
+export const addIngredient = (name) => {
+    return {
+        type: actionTypes.ADD_INGREDIENT,
+        ingredientName: name
+    }
+}
+
+export const removeIngredient = (name) => {
+    return {
+        type: actionTypes.REMOVE_INGREDIENT,
+        ingredientName: name
+    }
+}
