+++ src/store/actions/index.js
@@ -1,0 +1,6 @@
+export {
+    addIngredient,
+    removeIngredient
+} from './burgerBuilder';
+
+export {} from './order';
